03082023 PROCESOS INC

+++ SupplyChain/Client/HelperService/ProcunService.cs
@@ -12,7 +12,7 @@ using System.Timers;
 
 namespace SupplyChain.Client.HelperService
 {
-    public class ProcunService : BaseService<Procun, decimal>
+    public class ProcunService : BaseService<Procun, int>
     {
         private const string API = "api/Procun";
 
@@ -20,7 +20,7 @@ namespace SupplyChain.Client.HelperService
         {
         }
 
-        public async Task<bool> Existe(decimal id)
+        public async Task<bool> Existe(int id)
         {
             var response = await http.GetFromJsonAsync<bool>($"{API}/Existe/{id}");
             if (response.Error)
